data fetching from flipkart is done
